Fix include path not being found for header files
Although probably not an ideal method, I directly modified the
makefile to include the project relative "Inc" header folder
into the build. This worked, after unsuccessively trying to use the ECLIPSE
IDE to specify the include path for the directives.

I couldn't find the appropiate setting to modify so for now I'll stick
with modifying the makefile directly.