Fix bug - constant interrupt firing because in qsys
The interrupt for the switches was always firing whenever
the switch was in the ON position because it was configured as
a LEVEL IRQ inside qsys. I changed it to EDGE and now it works as expected.

Another issue I noticed in qsys ->
the buttons were configured as EDGE IRQ with FALLING edge detection. I changed
this to RISING edge detection, since the buttons are logic low on the dev board
but are then inverted before reaching the Nios processor.