Finish mode switching logic for switches ISR
Everything seems to work as expected -> when the config time mode is on,
the clock digits do not carry, and in every other mode they do. Furthermore,
the different display modes can be accessed and thus the user can switch between
seeing the current clock time, the current set alarm time, the current volume,
and the current song selected.

There is one small "bug" that seems to occur when a switch is slowly
set to the "on position". It seems that the switches interrupt fires multiple times
as indicated by the multiple printing of "The requested display is already active" printf
statement in the switches ISR. I'm suspecting that this is the cause of switch
debouncing, leading to multiple interrupt signals from the PIO.

In this particular case, the interrupt firing multiple times isn't really an issue. However,
a small delay can be added to the ISR to avoid this. Other techniques could also be used
such as a debouncing circuit instantiated on the FPGA fabric (for example).